+++ Sources/BasicDataSource.swift
@@ -8,7 +8,7 @@ import Foundation
 
 // MARK: - Model
 
-public struct BasicViewModel: ViewModel {
+public struct BasicDataSource: DataSource {
     public let title: String?
     public var sections: [Section]
 
@@ -23,7 +23,7 @@ public struct BasicViewModel: ViewModel {
     }
 
     public init(with data: Data, decoder: JSONDecoder = JSONDecoder()) throws {
-        self = try decoder.decode(BasicViewModel.self, from: data)
+        self = try decoder.decode(BasicDataSource.self, from: data)
     }
 }
 
@@ -41,36 +41,39 @@ public struct BasicSection: Section {
 
 public struct BasicItem: Item {
     public let identifier: String
-    public var model: Model
+    public var viewModel: ViewModel
+    public let child: DataSource?
 
-    public init(identifier: String, model: Model = BasicModel()) {
+    public init(identifier: String, viewModel: ViewModel = BasicViewModel(), child: DataSource? = nil) {
         self.identifier = identifier
-        self.model = model
+        self.viewModel = viewModel
+        self.child = child
     }
 
-    public init(identifier: String, title: String? = nil, detail: String? = nil, image: String? = nil) {
+    public init(identifier: String,
+                title: String? = nil, detail: String? = nil, image: String? = nil,
+                child: DataSource? = nil) {
         self.identifier = identifier
-        self.model = BasicModel(title: title, detail: detail, image: image)
+        self.viewModel = BasicViewModel(title: title, detail: detail, image: image)
+        self.child = child
     }
 }
 
-public struct BasicModel: Model {
+public struct BasicViewModel: ViewModel {
     public let title: String?
     public let detail: String?
     public let image: String?
-    public let child: ViewModel?
 
-    public init(title: String? = nil, detail: String? = nil, image: String? = nil, child: ViewModel? = nil) {
+    public init(title: String? = nil, detail: String? = nil, image: String? = nil) {
         self.title = title
         self.detail = detail
         self.image = image
-        self.child = child
     }
 }
 
 // MARK: - Codable
 
-extension BasicViewModel: Codable {
+extension BasicDataSource: Codable {
     public enum CodingKeys: String, CodingKey {
         case title, sections
     }
@@ -110,38 +113,38 @@ extension BasicSection: Codable {
 
 extension BasicItem: Codable {
     public enum CodingKeys: String, CodingKey {
-        case identifier, model
+        case identifier, viewModel, child
     }
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         identifier = try container.decode(String.self, forKey: .identifier)
-        model = try container.decode(BasicModel.self, forKey: .model)
+        viewModel = try container.decode(BasicViewModel.self, forKey: .viewModel)
+        child = try container.decodeIfPresent(BasicDataSource.self, forKey: .child)
     }
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(identifier, forKey: .identifier)
-        try container.encodeIfPresent(model as? BasicModel, forKey: .model)
+        try container.encodeIfPresent(viewModel as? BasicViewModel, forKey: .viewModel)
+        if let child = child as? BasicDataSource {
+            try container.encodeIfPresent(child, forKey: .child)
+        }
     }
 }
 
-extension BasicModel: Codable {
+extension BasicViewModel: Codable {
     public enum CodingKeys: String, CodingKey {
-        case title, detail, image, child
+        case title, detail, image
     }
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         title = try container.decodeIfPresent(String.self, forKey: .title)
         detail = try container.decodeIfPresent(String.self, forKey: .detail)
         image = try container.decodeIfPresent(String.self, forKey: .image)
-        child = try container.decodeIfPresent(BasicViewModel.self, forKey: .child)
     }
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encodeIfPresent(title, forKey: .title)
         try container.encodeIfPresent(detail, forKey: .detail)
         try container.encodeIfPresent(image, forKey: .image)
-        if let child = child as? BasicViewModel {
-            try container.encodeIfPresent(child, forKey: .child)
-        }
     }
 }